microgrid_base/requirements_docker_launch.txt
# pydantic
docker
# not "progressbar", which is not compatible
progressbar2
rich
schedule
better_exceptions
pytz